services/ml-service/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20

# ML/AI
torch==2.5.1
transformers==4.47.0
pillow==11.0.0

# Utilities
python-dotenv==1.0.1
